genai-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
PyMuPDF
pytesseract
Pillow
